src/polyvec/requirements.txt
boto3==1.38.3
botocore==1.38.3
certifi==2025.4.26
charset-normalizer==3.4.1
filelock==3.18.0
fsspec==2025.3.2
grpcio==1.72.0
grpcio-tools==1.72.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
pillow==11.2.1
protobuf==6.30.2
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.12.0
setuptools==78.1.1
six==1.17.0
sympy==1.13.3
torch==1.11.0
torchvision==0.12.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==1.26.20
wheel==0.45.1
